// libs/type.enum.ts
export enum TypeEnum {
  CHANNELS = 'channels',
  ASSETS = 'assets',
  FACETS = 'facets',
  FACET_VALUES = 'facetValues',
  COLLECTIONS = 'collections',
  PRODUCTS = 'products',
  PRODUCT_VARIANTS = 'productVariants',
  PRODUCT_OPTION_GROUPS = 'productOptionGroups',
  PRODUCT_OPTIONS = 'productOptions',
  COUNTRIES = 'countries',
  ZONES = 'zones',
  TAX_CATEGORIES = 'taxCategories',
  TAX_RATES = 'taxRates',
  PAYMENT_METHODS = 'paymentMethods',
  SHIPPING_METHODS = 'shippingMethods',
  ROLES = 'roles',
}